Remove CodeRefactor.Cotrols.LineEntryDialog
